+++ Assets/PotatoUtil/Scripts/Hash/FNVHash.cs
@@ -1,4 +1,4 @@
-﻿using BeauData;
+﻿using PotatoSerializer;
 using System;
 
 namespace PotatoUtil {
@@ -8,7 +8,7 @@ namespace PotatoUtil {
 	/// Fowler-Noll-Vo hashing algorthim.
 	/// http://www.isthe.com/chongo/tech/comp/fnv/
 	/// </summary>
-	public struct FNVHash : IEquatable<FNVHash>, IEquatable<int>, IEquatable<uint>, ISerializedProxy<uint> {
+	public struct FNVHash : IEquatable<FNVHash>, IEquatable<int>, IEquatable<uint>, ISerialProxy<uint> {
 
 		public static readonly FNVHash Empty = new FNVHash();
 
@@ -89,12 +89,12 @@ namespace PotatoUtil {
 			return HashLookup.ToString(this);
 		}
 
-		public uint GetProxyValue(ISerializerContext inContext) {
+		public uint GetSerialProxy() {
 			return m_hash;
 		}
 
-		public void SetProxyValue(uint inValue, ISerializerContext inContext) {
-			m_hash = inValue;
+		public void SetSerialProxy(uint value) {
+			m_hash = value;
 		}
 	}